+++ ethereum/weaksubjectivity/src/main/java/tech/pegasys/teku/weaksubjectivity/WeakSubjectivityCalculator.java
@@ -25,8 +25,10 @@ import tech.pegasys.teku.weaksubjectivity.config.WeakSubjectivityConfig;
 
 /**
  * This utility contains helpers for calculating weak-subjectivity-related values. Logic is derived
- * from: https://notes.ethereum.org/@adiasg/weak-subjectvity-eth2 and:
- * https://github.com/ethereum/eth2.0-specs/blob/weak-subjectivity-guide/specs/phase0/weak-subjectivity.md
+ * from <a href=https://notes.ethereum.org/@adiasg/weak-subjectvity-eth2>Weak Subjectivity in
+ * Eth2.0</a> and the <a
+ * href=https://github.com/ethereum/consensus-specs/blob/dev/specs/phase0/weak-subjectivity.md>Weak
+ * Subjectivity Guide</a>.
  */
 public class WeakSubjectivityCalculator {